refactor: moved ListenerState to entities package
