+++ operate/schema/src/main/java/io/camunda/operate/entities/ListenerState.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.operate;
+package io.camunda.operate.entities;
 
 public enum ListenerState {
   ACTIVE,
